requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.2
Brotli==1.0.9
bs4==0.0.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.4
cryptography==41.0.7
cssselect==1.2.0
discord==2.3.2
discord.py==2.3.2
fake-useragent==1.4.0
frozenlist==1.4.1
idna==3.4
importlib-metadata==7.0.1
lxml==5.1.0
mariadb==1.1.9
multidict==6.0.4
packaging==23.2
parse==1.20.0
pip==23.3.1
pycparser==2.21
pyee==8.2.2
pyOpenSSL==23.2.0
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
setuptools==68.2.2
soupsieve==2.5
tqdm==4.66.1
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
wheel==0.41.2
yarl==1.9.4
zipp==3.17.0
